Support fractional shift hours (e.g. 10.5)
shiftHours was stored as Int, truncating 10.5 to 10. Changed the
column to Float (DOUBLE PRECISION, applied to DB) and switched all
parseInt(shiftHours) calls to parseFloat across the employee API,
employee form (step 0.5), and payroll page so a 10.5-hour shift is
honored in hourly-rate and payroll calculations.

+++ app/(dashboard)/payroll/page.tsx
@@ -62,7 +62,7 @@ export default function PayrollPage() {
                 const autoAdvance = (Array.isArray(advances) ? advances : [])
                     .filter((a: any) => a.employeeId === e.id)
                     .reduce((s: number, a: any) => s + (a.amount ?? 0), 0)
-                const shiftHours = parseInt(e.shiftHours) || 8
+                const shiftHours = parseFloat(e.shiftHours) || 8
                 // Derive the TRUE hourly rate. basicSalary is a MONTHLY figure, so we
                 // must convert it — never multiply the monthly salary by hours directly.
                 //   • If a dailyRate is set (labour), hourly = dailyRate ÷ shift hours.
